AQ-335: Address reviewer comments, extract data setting of graph outside of the custom graph component
