+++ src/components/reusable_components/CustomGraph.js
@@ -9,60 +9,14 @@ import {
 import { Button } from "react-native-paper";
 import { THEME, LINE_GRAPH_THEME } from "../../constants/Theme";
 
-export default function CustomGraph() {
-  // Filler data until we connect to the backend
-  const mockData1 = [
-    { x: 1, y: 2 },
-    { x: 2, y: 3 },
-    { x: 3, y: 5 },
-    { x: 4, y: 4 },
-    { x: 5, y: 7 },
-  ];
-  const mockData2 = [
-    { x: 2, y: 6 },
-    { x: 3, y: 2 },
-    { x: 4, y: 9 },
-    { x: 6, y: 2 },
-    { x: 8, y: 1 },
-  ];
-
-  const mockData3 = [
-    { x: 0, y: 6 },
-    { x: 5, y: 10 },
-    { x: 6, y: 7 },
-    { x: 7, y: 9 },
-    { x: 8, y: 12 },
-  ];
-
-  const mockData4 = [
-    { x: 0, y: 6 },
-    { x: 5, y: 1 },
-    { x: 6, y: 7 },
-    { x: 7, y: 4 },
-    { x: 8, y: 10 },
-  ];
-
-  const [graphData, setGraphData] = useState(mockData1);
-  const [selectedTimeframe, setSelectedTimeframe] = useState(1);
-
-  // Update graphdata and change the selected timeframe
-  function handleTimeframeChange(timeframe) {
-    setSelectedTimeframe(timeframe);
-    switch (timeframe) {
-      case 1:
-        setGraphData(mockData1);
-        break;
-      case 2:
-        setGraphData(mockData2);
-        break;
-      case 3:
-        setGraphData(mockData3);
-        break;
-      case 4:
-        setGraphData(mockData4);
-        break;
-    }
-  }
+export default function CustomGraph(props) {
+  const {
+    graphData,
+    setGraphData,
+    selectedTimeframe,
+    setSelectedTimeframe,
+    handleTimeframeChange,
+  } = props;
 
   return (
     <View>
@@ -113,7 +67,9 @@ export default function CustomGraph() {
       <View style={styles.timeframeButtonOuterButtonRow}>
         <TouchableOpacity
           style={styles.timeframeButtonOuter}
-          onPress={() => handleTimeframeChange(1)}
+          onPress={() =>
+            handleTimeframeChange(1, setGraphData, setSelectedTimeframe)
+          }
         >
           <Button
             style={styles.timeframeButtonInner}
@@ -129,7 +85,9 @@ export default function CustomGraph() {
         </TouchableOpacity>
         <TouchableOpacity
           style={styles.timeframeButtonOuter}
-          onPress={() => handleTimeframeChange(2)}
+          onPress={() =>
+            handleTimeframeChange(2, setGraphData, setSelectedTimeframe)
+          }
         >
           <Button
             style={styles.timeframeButtonInner}
@@ -145,7 +103,9 @@ export default function CustomGraph() {
         </TouchableOpacity>
         <TouchableOpacity
           style={styles.timeframeButtonOuter}
-          onPress={() => handleTimeframeChange(3)}
+          onPress={() =>
+            handleTimeframeChange(3, setGraphData, setSelectedTimeframe)
+          }
         >
           <Button
             style={styles.timeframeButtonInner}
@@ -161,7 +121,9 @@ export default function CustomGraph() {
         </TouchableOpacity>
         <TouchableOpacity
           style={styles.timeframeButtonOuter}
-          onPress={() => handleTimeframeChange(4)}
+          onPress={() =>
+            handleTimeframeChange(4, setGraphData, setSelectedTimeframe)
+          }
         >
           <Button
             style={styles.timeframeButtonInner}
